Updated to .NET Standard 2 and new project system
A few things came for free, like the Linux support.
The whole build system was reworked as it all revolves around the
.NET CLI now.
Added Docker support to development and contribution.
Stylecop is not well supported for .NET Core, so that was removed
until support improves. The config files are still there.
Nuspec and package.config files were removed, .NET CLI does their
job now.
Tested against .NET Core and .NET Framework on Windows and on Debian.

Closes #62